12/17/2022 update League fragment
