+++ app/src/main/java/com/example/coolsports/presentation/league/LeagueFragment.kt
@@ -45,13 +45,9 @@ class LeagueFragment : BaseFragment() {
     private lateinit var navController: NavController
     private val viewModel by viewModels<LeagueViewModel>()
     var bundle = Bundle()
-    var leagueInfoList = ArrayList<LeagueData01>()
     var listLeague = mutableListOf<LeagueModel>()
-    var list = mutableListOf<LeagueModel>()
-    var listRules = ArrayList<LeagueData04>()
-    var leagueStandingGroup = ArrayList<LeagueStandingsGroupBase>()
 
-    var leagueId:Int = 0
+    var leagueId: Int = 0
 
 
     private val leagueListAdapter by lazy { LeagueListAdapter() }
@@ -65,86 +61,92 @@ class LeagueFragment : BaseFragment() {
     }
 
 
-
-
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.PREMIERE_LEAGUE),
-                Constants.PREMIERE_LEAGUE
+                Constants.PREMIERE_LEAGUE,"England","https://cdn1.77577cf.com/img-zq/Image/league_match/images/164577482086.png?win007=sell"
             )
         )
-        listLeague.add(LeagueModel(resources.getString(R.string.LA_LIGA), Constants.LA_LIGA))
-        listLeague.add(LeagueModel(resources.getString(R.string.SERIE_A), Constants.SERIE_A))
-        listLeague.add(LeagueModel(resources.getString(R.string.BUNDESLIGA), Constants.BUNDESLIGA))
-        listLeague.add(LeagueModel(resources.getString(R.string.LIGUE_1), Constants.LIGUE_1))
+        listLeague.add(LeagueModel(resources.getString(R.string.LA_LIGA), Constants.LA_LIGA,"Spain","https://cdn1.77577cf.com/img-zq/Image/league_match/images/20200415071847.jpg?win007=sell"))
+        listLeague.add(LeagueModel(resources.getString(R.string.SERIE_A), Constants.SERIE_A,"Italy","https://cdn1.77577cf.com/img-zq/Image/league_match/images/20190815175316.png?win007=sell"))
+        listLeague.add(LeagueModel(resources.getString(R.string.BUNDESLIGA), Constants.BUNDESLIGA,"Germany","https://cdn1.77577cf.com/img-zq/Image/league_match/images/20210630142615.png?win007=sell"))
+        listLeague.add(LeagueModel(resources.getString(R.string.LIGUE_1), Constants.LIGUE_1,"France","https://cdn1.77577cf.com/img-zq/Image/league_match/images/20210416104457.png?win007=sell"))
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.CHINESE_SUPER_LEAGUE),
-                Constants.CHINESE_SUPER_LEAGUE
+                Constants.CHINESE_SUPER_LEAGUE,
+                "China"
+                ,
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/20140111105821.jpg?win007=sell"
+
+            ,
             )
         )
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.AFC_CHAMPIONS_LEAGUE),
-                Constants.AFC_CHAMPIONS_LEAGUE
+                Constants.AFC_CHAMPIONS_LEAGUE,
+                "Asia",
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/20210104160334.png?win007=sell"
             )
         )
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.ASIAN_QUALIFIERS),
-                Constants.ASIAN_QUALIFIERS
+                Constants.ASIAN_QUALIFIERS,"International",
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/20210104160247.png?win007=sell"
             )
         )
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.SOUTH_AMERICAN_QUALIFIER),
-                Constants.SOUTH_AMERICAN_QUALIFIER
+                Constants.SOUTH_AMERICAN_QUALIFIER, "International","https://cdn1.77577cf.com/img-zq/Image/league_match/images/2008571027162493.jpg?win007=sell"
             )
         )
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.EUROPEAN_QUALIFIER),
-                Constants.EUROPEAN_QUALIFIER
+                Constants.EUROPEAN_QUALIFIER,
+                "International",
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/20085710192243008.jpg?win007=sell"
             )
         )
         listLeague.add(LeagueModel(resources.getString(R.string.WORLD_CUP), Constants.WORLD_CUP))
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.EUROPEAN_CUP),
-                Constants.EUROPEAN_CUP
+                Constants.EUROPEAN_CUP,
+                "International",
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/164885996047.png?win007=sell"
             )
         )
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.CONFEDERATIONS_CUP),
-                Constants.CONFEDERATIONS_CUP
+                Constants.CONFEDERATIONS_CUP,
+                "Europe",
+
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/20170531172156.jpg?win007=sell"
             )
         )
         listLeague.add(
             LeagueModel(
                 resources.getString(R.string.AMERICAS_CUP),
-                Constants.AMERICAS_CUP
+                Constants.AMERICAS_CUP,
+                "Americas",
+                "https://cdn1.77577cf.com/img-zq/Image/league_match/images/20200411095945.jpg?win007=sell"
             )
         )
+        leagueListAdapter.submitList(listLeague)
 
-        initObserver()
-
-        lifecycleScope.launch {
-            for (league in listLeague) {
-                viewModel.getLeagueInfo(league.leagueId!!, " ", 0)
-            }
-        }
     }
 
     @SuppressLint("ClickableViewAccessibility")
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         navController = view.findNavController()
         initRV()
-
-
 
     }
 
@@ -155,8 +157,6 @@ class LeagueFragment : BaseFragment() {
         leagueListAdapter.setItemTapListener(object : LeagueListAdapter.OnItemTap {
             override fun onTap(position: Int) {
                 leagueId = leagueListAdapter.getLeagueId(position)
-                Log.d(TAG, "leagueId========= $leagueId")
-
                 bundle.putInt("leagueId", leagueId)
                 if (findNavController().currentDestination?.id == R.id.LeagueFragment)
                     navController.navigate(
@@ -166,94 +166,7 @@ class LeagueFragment : BaseFragment() {
             }
         })
     }
+}
 
 
-    private fun initObserver() {
-        viewModel.mState.flowWithLifecycle(
-            this.lifecycle, Lifecycle.State.STARTED
-
-        ).onEach {
-            handleState(it)
-        }.launchIn(this.lifecycleScope)
-    }
-
-    private fun handleState(state: LeagueStateScreen) {
-        when (state) {
-            is LeagueStateScreen.IsLoading -> handleIsLoadingState(state.isLoading)
-            is LeagueStateScreen.Response -> handleLeagueResponse(state.leagueInfo)
-            is LeagueStateScreen.NoInternetException -> handleNetworkFailure(state.message)
-            is LeagueStateScreen.GeneralException -> handleException(state.message)
-            is LeagueStateScreen.StatusFailed -> handleFailure(state.message)
-            else -> {
-                Log.d(TAG, " no state run ")
-            }
-        }
-    }
-
-    private fun handleLeagueResponse(response: BaseLeagueInfo) {
-
-        leagueInfoList.addAll(response.leagueData01)
-
-        handleLeagueInfo(leagueInfoList)
-
-
-    }
-
-    private fun handleLeagueInfo(leagueInfoList: ArrayList<LeagueData01>) {
-        val sp = SPApp(requireContext())
-
-        for (league in leagueInfoList) {
-            list.add(
-                LeagueModel(
-                    leagueId = league.leagueId,
-                    country = league.countryEn,
-                    logo = league.leagueLogo,
-                    leagueName = league.nameEn
-                )
-            )
-            if (sp.language == Constants.SharedPreferenceKeys.CHINESE)
-                list.add(
-                    LeagueModel(
-                        leagueId = league.leagueId,
-                        country = league.countryCn,
-                        logo = league.leagueLogo,
-                        leagueName = league.nameCn
-                    )
-                )
-
-
-        }
-
-        leagueListAdapter.submitList(list.distinct())
-    }
-
-
-
-
-        private fun handleIsLoadingState(loading: Boolean) {
-            if (loading) {
-                showLoading()
-            } else {
-                hideLoading()
-            }
-        }
-
-        private fun handleFailure(message: String) {
-            showToast(message)
-            hideLoading()
-        }
-
-
-        private fun handleNetworkFailure(message: String) {
-            showToast(message)
-            hideLoading()
-        }
-
-        private fun handleException(message: String) {
-            showToast(message)
-            hideLoading()
-        }
-
-
-    }
