+++ app/src/main/java/com/example/coolsports/presentation/league/LeagueInfoFragment.kt
@@ -7,11 +7,22 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.flowWithLifecycle
+import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
 import com.example.coolsports.databinding.FragmentTeamInfoBinding
+import com.example.coolsports.domain.model.league.BaseLeagueInfo
+import com.example.coolsports.domain.model.league.LeagueData01
+import com.example.coolsports.domain.model.league.LeagueData04
+import com.example.coolsports.domain.model.league.LeagueModel
+import com.example.coolsports.domain.model.leagueStandings.LeagueStandingGroup.LeagueStandingsGroupBase
 import com.example.coolsports.presentation.base.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class LeagueInfoFragment : BaseFragment() {
@@ -21,6 +32,12 @@ class LeagueInfoFragment : BaseFragment() {
     private val binding get() = _binding!!
     private lateinit var navController: NavController
     private val viewModel by viewModels<LeagueViewModel>()
+    var leagueInfoList = ArrayList<LeagueData01>()
+    var listLeague = mutableListOf<LeagueModel>()
+    var list = mutableListOf<LeagueModel>()
+    var listRules = ArrayList<LeagueData04>()
+    var leagueStandingGroup = ArrayList<LeagueStandingsGroupBase>()
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -34,12 +51,19 @@ class LeagueInfoFragment : BaseFragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         navController = view.findNavController()
 
-        goToNext()
+        initObserver()
 
         val leagueId = arguments!!.getInt("LeagueId")
         Log.d(TAG, leagueId.toString())
 
+
+
+        lifecycleScope.launch {
+            viewModel.getLeagueInfo(leagueId, " ", 0)
+
+        }
     }
+
 
 
 //        private fun testCasting(leagueInfoBase: BaseLeagueInfo): BaseLeagueInfo {
@@ -63,19 +87,68 @@ class LeagueInfoFragment : BaseFragment() {
 //        }
 
 
-    fun goToNext() {
-//        binding.goTeamBtn.setOnClickListener {
-//            if (findNavController().currentDestination?.id == R.id.LeagueFragment)
-//                navController.navigate(R.id.action_leagueFragment_to_teamStandingsFragment,bundle)
-//        }
-//        binding.goPlayerBtn.setOnClickListener {
-//            if (findNavController().currentDestination?.id == R.id.LeagueFragment)
-//                navController.navigate(R.id.action_leagueFragment_to_playerStandingsFragment)
-//        }
 
+
+    private fun initObserver() {
+        viewModel.mState.flowWithLifecycle(
+            this.lifecycle, Lifecycle.State.STARTED
+
+        ).onEach {
+            handleState(it)
+        }.launchIn(this.lifecycleScope)
+    }
+
+    private fun handleState(state: LeagueStateScreen) {
+        when (state) {
+            is LeagueStateScreen.IsLoading -> handleIsLoadingState(state.isLoading)
+            is LeagueStateScreen.Response -> handleLeagueResponse(state.leagueInfo)
+            is LeagueStateScreen.NoInternetException -> handleNetworkFailure(state.message)
+            is LeagueStateScreen.GeneralException -> handleException(state.message)
+            is LeagueStateScreen.StatusFailed -> handleFailure(state.message)
+            else -> {
+                Log.d(TAG, " no state run ")
+            }
+        }
+    }
+
+    private fun handleLeagueResponse(response: BaseLeagueInfo) {
+        leagueInfoList.addAll(response.leagueData01)
 
     }
+
+
+
+
+
+
+    private fun handleIsLoadingState(loading: Boolean) {
+        if (loading) {
+            showLoading()
+        } else {
+            hideLoading()
+        }
+    }
+
+    private fun handleFailure(message: String) {
+        showToast(message)
+        hideLoading()
+    }
+
+
+    private fun handleNetworkFailure(message: String) {
+        showToast(message)
+        hideLoading()
+    }
+
+    private fun handleException(message: String) {
+        showToast(message)
+        hideLoading()
+    }
+
+
 }
+
+
